requirements.txt
django
apscheduler
django-apscheduler
django-cleanup
django-crispy-forms
django-hcaptcha
gunicorn
mysql
mysqlclient
